Split navigator tests into multiple files
